Create FourRelayActuatorWithButtonTempLCD.ino
This sketch's default set-up is for 4 relays, an actuator button to activate relay of choice, temperature sensors and a 20x4 LCD. The relay button is set up to use the internal pull-up resistor, so no resistor is needed.  Just connect pin to push button and then to ground.  You can change the number of relays and possibly add more if you use the analog pins. You can use any Hitachi character LCD.  Just change the rows and columns to match.